pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sgprotonet"
version = "0.1.0"
description = "Semantic Guided Prototypical Networks for few-shot medical image classification"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.30",
    "open_clip_torch>=2.20",
    "omegaconf>=2.3",
    "wandb>=0.15",
    "pandas>=2.0",
    "nltk>=3.8",
    "scipy>=1.10",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
